Break up plotting module
