+++ pyvista/plotting/__init__.py
@@ -1,4 +1,7 @@
+from .theme import *
+from .tools import *
 from .plotting import *
+from .helpers import *
 from .colors import *
 from .export_vtkjs import export_plotter_vtkjs, get_vtkjs_url
 from .ipy_tools import (Clip, Isocontour, ManySlicesAlongAxis,
