+++ pyvista/plotting/plotting.py
@@ -2,11 +2,9 @@
 pyvista plotting module
 """
 import collections
-import ctypes
 import logging
 import os
 import time
-from subprocess import PIPE, Popen
 from threading import Thread
 
 import imageio
@@ -20,6 +18,8 @@ from pyvista.utilities import (convert_array, get_scalar, is_pyvista_obj,
 
 from .colors import get_cmap_safe
 from .export_vtkjs import export_plotter_vtkjs
+from .theme import *
+from .tools import *
 
 _ALL_PLOTTERS = {}
 
@@ -30,309 +30,10 @@ def close_all():
     _ALL_PLOTTERS.clear()
     return True
 
-MAX_N_COLOR_BARS = 10
-PV_BACKGROUND = [82/255., 87/255., 110/255.]
-FONT_KEYS = {'arial': vtk.VTK_ARIAL,
-             'courier': vtk.VTK_COURIER,
-             'times': vtk.VTK_TIMES}
-
 
 log = logging.getLogger(__name__)
 log.setLevel('CRITICAL')
 
-
-rcParams = {
-    'background' : [0.3, 0.3, 0.3],
-    'camera' : {
-        'position' : [1, 1, 1],
-        'viewup' : [0, 0, 1],
-    },
-    'window_size' : [1024, 768],
-    'font' : {
-        'family' : 'courier',
-        'size' : 12,
-        'title_size': None,
-        'label_size' : None,
-        'color' : [1, 1, 1],
-        'fmt' : None,
-    },
-    'cmap' : 'viridis',
-    'color' : 'white',
-    'nan_color' : 'darkgray',
-    'edge_color' : 'black',
-    'outline_color' : 'white',
-    'colorbar_orientation' : 'horizontal',
-    'colorbar_horizontal' : {
-        'width' : 0.60,
-        'height' : 0.08,
-        'position_x' : 0.35,
-        'position_y' : 0.02,
-    },
-    'colorbar_vertical' : {
-        'width' : 0.1,
-        'height' : 0.8,
-        'position_x' : 0.85,
-        'position_y' : 0.1,
-    },
-    'show_scalar_bar' : True,
-    'show_edges' : False,
-    'lighting' : True,
-    'interactive' : False,
-    'render_points_as_spheres' : False,
-    'use_panel' : True,
-    'transparent_background' : False
-}
-
-DEFAULT_THEME = dict(rcParams)
-
-def set_plot_theme(theme):
-    """Set the plotting parameters to a predefined theme"""
-    if theme.lower() in ['paraview', 'pv']:
-        rcParams['background'] = PV_BACKGROUND
-        rcParams['cmap'] = 'coolwarm'
-        rcParams['font']['family'] = 'arial'
-        rcParams['font']['label_size'] = 16
-        rcParams['show_edges'] = False
-    elif theme.lower() in ['document', 'doc', 'paper', 'report']:
-        rcParams['background'] = 'white'
-        rcParams['cmap'] = 'viridis'
-        rcParams['font']['size'] = 18
-        rcParams['font']['title_size'] = 18
-        rcParams['font']['label_size'] = 18
-        rcParams['font']['color'] = 'black'
-        rcParams['show_edges'] = False
-        rcParams['color'] = 'tan'
-        rcParams['outline_color'] = 'black'
-    elif theme.lower() in ['night', 'dark']:
-        rcParams['background'] = 'black'
-        rcParams['cmap'] = 'viridis'
-        rcParams['font']['color'] = 'white'
-        rcParams['show_edges'] = False
-        rcParams['color'] = 'tan'
-        rcParams['outline_color'] = 'white'
-        rcParams['edge_color'] = 'white'
-    elif theme.lower() in ['default']:
-        for k,v in DEFAULT_THEME.items():
-            rcParams[k] = v
-
-
-def run_from_ipython():
-    """ returns True when run from IPython """
-    try:
-        py = __IPYTHON__
-        return True
-    except NameError:
-        return False
-
-
-
-def opacity_transfer_function(key, n_colors):
-    """Get the opacity transfer function results: range from 0 to 255.
-    """
-    sigmoid = lambda x: np.array(1 / (1 + np.exp(-x)) * 255, dtype=np.uint8)
-    transfer_func = {
-        'linear': np.linspace(0, 255, n_colors, dtype=np.uint8),
-        'linear_r': np.linspace(0, 255, n_colors, dtype=np.uint8)[::-1],
-        'geom': np.geomspace(1e-6, 255, n_colors, dtype=np.uint8),
-        'geom_r': np.geomspace(255, 1e-6, n_colors, dtype=np.uint8),
-        'sigmoid': sigmoid(np.linspace(-10.,10., n_colors)),
-        'sigmoid_3': sigmoid(np.linspace(-3.,3., n_colors)),
-        'sigmoid_4': sigmoid(np.linspace(-4.,4., n_colors)),
-        'sigmoid_5': sigmoid(np.linspace(-5.,5., n_colors)),
-        'sigmoid_6': sigmoid(np.linspace(-6.,6., n_colors)),
-        'sigmoid_7': sigmoid(np.linspace(-7.,7., n_colors)),
-        'sigmoid_8': sigmoid(np.linspace(-8.,8., n_colors)),
-        'sigmoid_9': sigmoid(np.linspace(-9.,9., n_colors)),
-        'sigmoid_10': sigmoid(np.linspace(-10.,10., n_colors)),
-
-    }
-    try:
-        return transfer_func[key]
-    except KeyError:
-        raise KeyError('opactiy transfer function ({}) unknown.'.format(key))
-
-
-def plot(var_item, off_screen=None, full_screen=False, screenshot=None,
-         interactive=True, cpos=None, window_size=None,
-         show_bounds=False, show_axes=True, notebook=None, background=None,
-         text='', return_img=False, eye_dome_lighting=False, use_panel=None,
-         volume=False, **kwargs):
-    """
-    Convenience plotting function for a vtk or numpy object.
-
-    Parameters
-    ----------
-    item : vtk or numpy object
-        VTK object or numpy array to be plotted.
-
-    off_screen : bool
-        Plots off screen when True.  Helpful for saving screenshots
-        without a window popping up.
-
-    full_screen : bool, optional
-        Opens window in full screen.  When enabled, ignores window_size.
-        Default False.
-
-    screenshot : str or bool, optional
-        Saves screenshot to file when enabled.  See:
-        help(pyvistanterface.Plotter.screenshot).  Default disabled.
-
-        When True, takes screenshot and returns numpy array of image.
-
-    window_size : list, optional
-        Window size in pixels.  Defaults to [1024, 768]
-
-    show_bounds : bool, optional
-        Shows mesh bounds when True.  Default False. Alias ``show_grid`` also
-        accepted.
-
-    notebook : bool, optional
-        When True, the resulting plot is placed inline a jupyter notebook.
-        Assumes a jupyter console is active.
-
-    show_axes : bool, optional
-        Shows a vtk axes widget.  Enabled by default.
-
-    text : str, optional
-        Adds text at the bottom of the plot.
-
-    volume : bool, optional
-        Use the ``add_volume`` method for volume rendering.
-
-    **kwargs : optional keyword arguments
-        See help(Plotter.add_mesh) for additional options.
-
-    Returns
-    -------
-    cpos : list
-        List of camera position, focal point, and view up.
-
-    img :  numpy.ndarray
-        Array containing pixel RGB and alpha.  Sized:
-        [Window height x Window width x 3] for transparent_background=False
-        [Window height x Window width x 4] for transparent_background=True
-        Returned only when screenshot enabled
-
-    """
-    if notebook is None:
-        if run_from_ipython():
-            try:
-                notebook = type(get_ipython()).__module__.startswith('ipykernel.')
-            except NameError:
-                pass
-
-    if notebook:
-        off_screen = notebook
-    plotter = Plotter(off_screen=off_screen, notebook=notebook)
-    if show_axes:
-        plotter.add_axes()
-
-    plotter.set_background(background)
-
-    if isinstance(var_item, list):
-        if len(var_item) == 2:  # might be arrows
-            isarr_0 = isinstance(var_item[0], np.ndarray)
-            isarr_1 = isinstance(var_item[1], np.ndarray)
-            if isarr_0 and isarr_1:
-                plotter.add_arrows(var_item[0], var_item[1])
-            else:
-                for item in var_item:
-                    if volume or (isinstance(item, np.ndarray) and item.ndim == 3):
-                        plotter.add_volume(item, **kwargs)
-                    else:
-                        plotter.add_mesh(item, **kwargs)
-        else:
-            for item in var_item:
-                if volume or (isinstance(item, np.ndarray) and item.ndim == 3):
-                    plotter.add_volume(item, **kwargs)
-                else:
-                    plotter.add_mesh(item, **kwargs)
-    else:
-        if volume or (isinstance(var_item, np.ndarray) and var_item.ndim == 3):
-            plotter.add_volume(var_item, **kwargs)
-        else:
-            plotter.add_mesh(var_item, **kwargs)
-
-    if text:
-        plotter.add_text(text)
-
-    if show_bounds or kwargs.get('show_grid', False):
-        if kwargs.get('show_grid', False):
-            plotter.show_grid()
-        else:
-            plotter.show_bounds()
-
-    if cpos is None:
-        cpos = plotter.get_default_cam_pos()
-        plotter.camera_position = cpos
-        plotter.camera_set = False
-    else:
-        plotter.camera_position = cpos
-
-    if eye_dome_lighting:
-        plotter.enable_eye_dome_lighting()
-
-    result = plotter.show(window_size=window_size,
-                          auto_close=False,
-                          interactive=interactive,
-                          full_screen=full_screen,
-                          screenshot=screenshot,
-                          return_img=return_img,
-                          use_panel=use_panel)
-
-    # close and return camera position and maybe image
-    plotter.close()
-
-    # Result will be handled by plotter.show(): cpos or [cpos, img]
-    return result
-
-
-def plot_arrows(cent, direction, **kwargs):
-    """
-    Plots arrows as vectors
-
-    Parameters
-    ----------
-    cent : np.ndarray
-        Accepts a single 3d point or array of 3d points.
-
-    directions : np.ndarray
-        Accepts a single 3d point or array of 3d vectors.
-        Must contain the same number of items as cent.
-
-    **kwargs : additional arguments, optional
-        See help(pyvista.Plot)
-
-    Returns
-    -------
-    Same as Plot.  See help(pyvista.Plot)
-
-    """
-    return plot([cent, direction], **kwargs)
-
-
-def system_supports_plotting():
-    """
-    Check if x server is running
-
-    Returns
-    -------
-    system_supports_plotting : bool
-        True when on Linux and running an xserver.  Returns None when
-        on a non-linux platform.
-
-    """
-    try:
-        if os.environ['ALLOW_PLOTTING'].lower() == 'true':
-            return True
-    except KeyError:
-        pass
-    try:
-        p = Popen(["xset", "-q"], stdout=PIPE, stderr=PIPE)
-        p.communicate()
-        return p.returncode == 0
-    except:
-        return False
 
 
 class BasePlotter(object):
@@ -3378,134 +3079,3 @@ class Plotter(BasePlotter):
     def render(self):
         """ renders main window """
         self.ren_win.Render()
-
-
-
-def single_triangle():
-    """ A single PolyData triangle """
-    points = np.zeros((3, 3))
-    points[1] = [1, 0, 0]
-    points[2] = [0.5, 0.707, 0]
-    cells = np.array([[3, 0, 1, 2]], ctypes.c_long)
-    return pyvista.PolyData(points, cells)
-
-
-def parse_color(color):
-    """ Parses color into a vtk friendly rgb list """
-    if color is None:
-        color = rcParams['color']
-    if isinstance(color, str):
-        return pyvista.string_to_rgb(color)
-    elif len(color) == 3:
-        return color
-    else:
-        raise Exception("""
-    Invalid color input
-    Must ba string, rgb list, or hex color string.  For example:
-        color='white'
-        color='w'
-        color=[1, 1, 1]
-        color='#FFFFFF'""")
-
-
-def parse_font_family(font_family):
-    """ checks font name """
-    # check font name
-    font_family = font_family.lower()
-    if font_family not in ['courier', 'times', 'arial']:
-        raise Exception('Font must be either "courier", "times" ' +
-                        'or "arial"')
-
-    return FONT_KEYS[font_family]
-
-
-def plot_compare_four(data_a, data_b, data_c, data_d, disply_kwargs=None,
-                      plotter_kwargs=None, show_kwargs=None, screenshot=None,
-                      camera_position=None, outline=None, outline_color='k',
-                      labels=('A', 'B', 'C', 'D')):
-    """Plot a 2 by 2 comparison of data objects. Plotting parameters and camera
-    positions will all be the same.
-    """
-    datasets = [[data_a, data_b], [data_c, data_d]]
-    labels = [labels[0:2], labels[2:4]]
-
-    if plotter_kwargs is None:
-        plotter_kwargs = {}
-    if disply_kwargs is None:
-        disply_kwargs = {}
-    if show_kwargs is None:
-        show_kwargs = {}
-
-    p = pyvista.Plotter(shape=(2,2), **plotter_kwargs)
-
-    for i in range(2):
-        for j in range(2):
-            p.subplot(i, j)
-            p.add_mesh(datasets[i][j], **disply_kwargs)
-            p.add_text(labels[i][j])
-            if is_pyvista_obj(outline):
-                p.add_mesh(outline, color=outline_color)
-            if camera_position is not None:
-                p.camera_position = camera_position
-
-    return p.show(screenshot=screenshot, **show_kwargs)
-
-
-def create_axes_orientation_box(line_width=1, text_scale=0.366667,
-                                edge_color='black', x_color='lightblue',
-                                y_color='seagreen', z_color='tomato',
-                                x_face_color=(255, 0, 0),
-                                y_face_color=(0, 255, 0),
-                                z_face_color=(0, 0, 255),
-                                color_box=False):
-    """Create a Box axes orientation widget with labels.
-    """
-    axes_actor = vtk.vtkAnnotatedCubeActor()
-    axes_actor.SetFaceTextScale(text_scale)
-    axes_actor.SetXPlusFaceText("X+")
-    axes_actor.SetXMinusFaceText("X-")
-    axes_actor.SetYPlusFaceText("Y+")
-    axes_actor.SetYMinusFaceText("Y-")
-    axes_actor.SetZPlusFaceText("Z+")
-    axes_actor.SetZMinusFaceText("Z-")
-    axes_actor.GetTextEdgesProperty().SetColor(parse_color(edge_color))
-    axes_actor.GetTextEdgesProperty().SetLineWidth(line_width)
-    axes_actor.GetXPlusFaceProperty().SetColor(parse_color(x_color))
-    axes_actor.GetXMinusFaceProperty().SetColor(parse_color(x_color))
-    axes_actor.GetYPlusFaceProperty().SetColor(parse_color(y_color))
-    axes_actor.GetYMinusFaceProperty().SetColor(parse_color(y_color))
-    axes_actor.GetZPlusFaceProperty().SetColor(parse_color(z_color))
-    axes_actor.GetZMinusFaceProperty().SetColor(parse_color(z_color))
-
-    if color_box:
-        # Hide the cube so we can color each face
-        axes_actor.GetCubeProperty().SetOpacity(0)
-
-        cube = pyvista.Cube()
-        face_colors = np.array([x_face_color,
-                                x_face_color,
-                                y_face_color,
-                                y_face_color,
-                                z_face_color,
-                                z_face_color,
-                               ], dtype=np.uint8)
-
-        cube.cell_arrays['face_colors'] = face_colors
-
-        cube_mapper = vtk.vtkPolyDataMapper()
-        cube_mapper.SetInputData(cube)
-        cube_mapper.SetColorModeToDirectScalars()
-        cube_mapper.Update()
-
-        cube_actor = vtk.vtkActor()
-        cube_actor.SetMapper(cube_mapper)
-        cube_actor.GetProperty().BackfaceCullingOn()
-
-        prop_assembly = vtk.vtkPropAssembly()
-        prop_assembly.AddPart(axes_actor)
-        prop_assembly.AddPart(cube_actor)
-        actor = prop_assembly
-    else:
-        actor = axes_actor
-
-    return actor
